Updates: Code almost functioning. Only need to add code for returning tempDir to SBH
