+++ app.py
@@ -2,6 +2,7 @@ from flask import Flask, request, send_file
 from flask.helpers import make_response
 from werkzeug.utils import secure_filename
 import os, zipfile, tempfile
+import sys
 import lib
 
 app = Flask(__name__)
@@ -36,26 +37,38 @@ def evaluate():
 @app.route('/run', methods=['POST'])
 def run():
 
-        # temporary directory to write intermediate files to
-        #temp_dir = tempfile.TemporaryDirectory()
+        #temporary directory to write intermediate files to
+        temp_dir = tempfile.TemporaryDirectory()
 
-        #data = request.get_json(force=True)
+        data = request.get_json(force=True)
 
-        #top_level_url = data['top_level']
-        #complete_sbol = data['complete_sbol']
-        #instance_url = data['instanceUrl']
-        #genbank_url = data['genbank']
-        #size = data['size']
-        #rdf_type = data['type']
-        #shallow_sbol = data['shallow_sbol']
-
-        #url = complete_sbol.replace('/sbol', '')
+        top_level_url = data['top_level']
+        complete_sbol = data['complete_sbol']
+        instance_url = data['instanceUrl']
+        genbank_url = data['genbank']
+        size = data['size']
+        rdf_type = data['type']
+        shallow_sbol = data['shallow_sbol']
+        url = complete_sbol.replace('/sbol', '')
 
         try:
             # ~~~~~~~~~~~~~ REPLACE THIS SECTION WITH OWN RUN CODE ~~~~~~~~~~~~~~~
             archive_url = 'https://subtest.synbiohub.org/download/ECEN5003_ToggleSwitch_LukasBuecherl.omex' #top_level_url + '/archive'
-            return lib.call(archive_url)
+            image = lib.call(archive_url)
+
+            out_name = "iBioSim_Results.png"
+            file_out_name = os.path.join(temp_dir.name, out_name)
+
+            with open(file_out_name, 'wb') as out_file:
+                out_file.write(image)
+
+
             # ~~~~~~~~~~~~~~~~~~~~~~~~~~~ END SECTION ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+            return send_from_directory(temp_dir.name, out_name,
+                                    as_attachment=True,
+                                    attachment_filename=out_name)
+
         except Exception as e:
             exc_type, exc_obj, exc_tb = sys.exc_info()
             fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
