+++ jsbml.log
@@ -11,3 +11,26 @@
 2021-11-11 13:10:25,125 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
 2021-11-11 13:16:04,958 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
 2021-11-11 13:16:24,127 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:26:29,580 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:29:23,860 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:30:10,412 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:31:55,281 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:32:58,446 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:34:41,190 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:34:51,946 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:36:43,281 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:37:14,130 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:39:22,357 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:39:56,466 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:40:48,500 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:41:40,910 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:44:27,381 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:44:51,061 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:45:14,346 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:48:20,363 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:49:13,776 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:49:39,232 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:49:58,417 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:50:08,375 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:50:34,647 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-12 17:51:58,220 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
